Extract Subject model to new app
