+++ applications/models.py
@@ -1,12 +1,6 @@
 from django.db import models
 
-
-class Subject(models.Model):
-    name = models.CharField(max_length=255)
-    slug = models.SlugField(unique=True)
-
-    def __str__(self) -> str:  # type: ignore[override]
-        return self.name
+from subjects.models import Subject
 
 
 class Application(models.Model):
