+++ fractalschool/settings.py
@@ -48,6 +48,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
+    'subjects',
     'accounts',
     'apps.recsys',
     'applications',
