+++ subjects/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from .models import Subject
+
+
+@admin.register(Subject)
+class SubjectAdmin(admin.ModelAdmin):
+    prepopulated_fields = {"slug": ("name",)}
+    list_display = ("name",)
